+++ be/src/main/java/codesquad/kr/gyeonggidoidle/issuetracker/config/FilterConfig.java
@@ -1,6 +1,7 @@
 package codesquad.kr.gyeonggidoidle.issuetracker.config;
 
-import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.JwtAuthorizationFilter;
+import codesquad.kr.gyeonggidoidle.issuetracker.config.filter.CorsFilter;
+import codesquad.kr.gyeonggidoidle.issuetracker.config.filter.JwtAuthorizationFilter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import javax.servlet.Filter;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
