+++ be/src/main/java/codesquad/kr/gyeonggidoidle/issuetracker/config/OauthConfig.java
@@ -1,7 +1,7 @@
 package codesquad.kr.gyeonggidoidle.issuetracker.config;
 
-import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.entity.OauthAdapter;
-import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.entity.OauthProperties;
+import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.OauthAdapter;
+import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.OauthProperties;
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.entity.OauthProvider;
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.repository.InMemoryProviderRepository;
 import java.util.Map;
